requirements.txt
certifi==2018.4.16
dj-database-url==0.5.0
Django==2.0.6
django-heroku==0.3.1
djangorestframework==3.8.2
psycopg2-binary==2.7.5
pytz==2018.4
whitenoise==3.3.1
